requirements.txt
tld
fuzzywuzzy
requests

# Configuration file support (optional, recommended)
PyYAML>=5.4

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
